fix: controller status
